requirements.txt
# Core dependencies
requests>=2.32.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Content extraction
youtube-transcript-api>=1.0.0
beautifulsoup4>=4.12.0

# Optional: For PDF processing
# PyPDF2>=3.0.0

# Optional: For advanced LLM providers
# openai>=1.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# mypy>=1.0.0
